ref(js): Rename constants/{globalSelectionHeader -> pageFilters}
