+++ static/app/components/organizations/multipleEnvironmentSelector.tsx
@@ -12,7 +12,7 @@ import Highlight from 'sentry/components/highlight';
 import GlobalSelectionHeaderRow from 'sentry/components/organizations/globalSelectionHeaderRow';
 import HeaderItem from 'sentry/components/organizations/headerItem';
 import MultipleSelectorSubmitRow from 'sentry/components/organizations/multipleSelectorSubmitRow';
-import {ALL_ACCESS_PROJECTS} from 'sentry/constants/globalSelectionHeader';
+import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import {IconWindow} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import ConfigStore from 'sentry/stores/configStore';
