+++ static/app/components/quickTrace/utils.tsx
@@ -1,7 +1,7 @@
 import {Location, LocationDescriptor} from 'history';
 
 import {getParams} from 'sentry/components/organizations/globalSelectionHeader/getParams';
-import {ALL_ACCESS_PROJECTS} from 'sentry/constants/globalSelectionHeader';
+import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import {OrganizationSummary} from 'sentry/types';
 import {Event} from 'sentry/types/event';
 import {defined} from 'sentry/utils';
