+++ static/app/utils/performance/quickTrace/utils.tsx
@@ -2,7 +2,7 @@ import omit from 'lodash/omit';
 import moment from 'moment-timezone';
 
 import {getTraceDateTimeRange} from 'sentry/components/events/interfaces/spans/utils';
-import {ALL_ACCESS_PROJECTS} from 'sentry/constants/globalSelectionHeader';
+import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import {OrganizationSummary} from 'sentry/types';
 import {Event, EventTransaction} from 'sentry/types/event';
 import {trackAnalyticsEvent} from 'sentry/utils/analytics';
