+++ static/app/views/dashboardsV2/widget/metricWidget/statsRequest.tsx
@@ -6,7 +6,7 @@ import pick from 'lodash/pick';
 import {addErrorMessage} from 'sentry/actionCreators/indicator';
 import {Client} from 'sentry/api';
 import {getParams} from 'sentry/components/organizations/globalSelectionHeader/getParams';
-import {URL_PARAM} from 'sentry/constants/globalSelectionHeader';
+import {URL_PARAM} from 'sentry/constants/pageFilters';
 import {t} from 'sentry/locale';
 import {
   GlobalSelection,
