+++ static/app/views/performance/traceDetails/index.tsx
@@ -6,7 +6,7 @@ import {Client} from 'sentry/api';
 import NoProjectMessage from 'sentry/components/noProjectMessage';
 import {getParams} from 'sentry/components/organizations/globalSelectionHeader/getParams';
 import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
-import {ALL_ACCESS_PROJECTS} from 'sentry/constants/globalSelectionHeader';
+import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import {t} from 'sentry/locale';
 import {PageContent} from 'sentry/styles/organization';
 import {Organization} from 'sentry/types';
