+++ static/app/views/releases/detail/header/releaseHeader.tsx
@@ -13,7 +13,7 @@ import ListLink from 'sentry/components/links/listLink';
 import NavTabs from 'sentry/components/navTabs';
 import Tooltip from 'sentry/components/tooltip';
 import Version from 'sentry/components/version';
-import {URL_PARAM} from 'sentry/constants/globalSelectionHeader';
+import {URL_PARAM} from 'sentry/constants/pageFilters';
 import {IconCopy, IconOpen} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import space from 'sentry/styles/space';
