+++ static/app/views/releases/list/index.tsx
@@ -20,7 +20,7 @@ import SearchBar from 'sentry/components/searchBar';
 import SmartSearchBar from 'sentry/components/smartSearchBar';
 import {ItemType} from 'sentry/components/smartSearchBar/types';
 import {DEFAULT_STATS_PERIOD} from 'sentry/constants';
-import {ALL_ACCESS_PROJECTS} from 'sentry/constants/globalSelectionHeader';
+import {ALL_ACCESS_PROJECTS} from 'sentry/constants/pageFilters';
 import {releaseHealth} from 'sentry/data/platformCategories';
 import {IconInfo} from 'sentry/icons';
 import {t} from 'sentry/locale';
