+++ static/app/views/userFeedback/utils.tsx
@@ -1,7 +1,7 @@
 import pick from 'lodash/pick';
 import * as qs from 'query-string';
 
-import {URL_PARAM} from 'sentry/constants/globalSelectionHeader';
+import {URL_PARAM} from 'sentry/constants/pageFilters';
 
 const DEFAULT_STATUS = 'unresolved';
